worker-service/requirements.txt
psycopg2-binary==2.9.6
redis==4.5.5
celery==5.2.7
python-dotenv==1.0.0